fix(subsample): PIVOT/UNION propagation + cadence integer-overflow hardening
- PIVOT silently dropped SUBSAMPLE: rewritePivot moved LIMIT/ORDER BY off the
  pivot model but not SUBSAMPLE, so it was discarded. Add moveSubsampleFrom.
- SUBSAMPLE on a non-final UNION arm was applied silently (first-arm-only);
  reject it in SqlParser like ORDER BY / LIMIT (+ regression test).
- cadence: 'stride + offset' used int arithmetic and could overflow to a
  negative buffer index (OOB native read); use long. computeCadenceOffset used
  Math.abs(h) % stride (negative for the seed hashing to Long.MIN_VALUE); use
  Math.floorMod. Mirror the fix in the test helper.

SubsampleTest 140/140 green.

+++ core/src/main/java/io/questdb/griffin/engine/table/CadenceAlgorithm.java
@@ -69,7 +69,10 @@ class CadenceAlgorithm {
         }
         int lastIdx = bufferSize - 1;
         // Emit stride-selected rows: stride + offset, 2*stride + offset, ...
-        for (int pos = stride + offset; pos < bufferSize; pos += stride) {
+        // Use long for the running index: stride + offset can exceed Integer.MAX_VALUE
+        // (stride up to Integer.MAX_VALUE, offset in [0, stride)), which would wrap to a
+        // negative int and produce an out-of-bounds native buffer read downstream.
+        for (long pos = (long) stride + offset; pos < bufferSize; pos += stride) {
             if ((pos & 0xFFF) == 0) {
                 circuitBreaker.statefulThrowExceptionIfTripped();
             }
